blossom_garden/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django
django-cors-headers==4.3.1
djangorestframework==3.14.0
idna==3.6
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
pytz==2024.1
requests==2.31.0
sqlparse
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
